reorganiza estrutura do backend movendo arquivos para /src

+++ backend/src/controllers/investimentoController.ts
@@ -1,6 +1,9 @@
-// src/controllers/investimentoController.ts
 import { Request, Response } from 'express';
+import mongoose from 'mongoose';
 import Investimento from '../models/Investimento';
+
+// Função para validar se o ID é um ObjectId válido
+const isValidObjectId = (id: string) => mongoose.Types.ObjectId.isValid(id);
 
 export const getInvestimentos = async (req: Request, res: Response) => {
   try {
@@ -12,10 +15,10 @@ export const getInvestimentos = async (req: Request, res: Response) => {
 };
 
 export const addInvestimento = async (req: Request, res: Response) => {
-  const { nome, tipo, valor } = req.body;
+  const { nome, tipo, valor, data } = req.body;
 
   try {
-    const novoInvestimento = new Investimento({ nome, tipo, valor });
+    const novoInvestimento = new Investimento({ nome, tipo, valor, data });
     await novoInvestimento.save();
     res.status(201).json(novoInvestimento);
   } catch (error) {
@@ -25,14 +28,24 @@ export const addInvestimento = async (req: Request, res: Response) => {
 
 export const updateInvestimento = async (req: Request, res: Response) => {
   const { id } = req.params;
-  const { nome, tipo, valor } = req.body;
+  const { nome, tipo, valor, data } = req.body;
+
+  // Valida se o ID é um ObjectId válido
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ message: 'ID inválido' });
+  }
 
   try {
     const investimentoAtualizado = await Investimento.findByIdAndUpdate(
       id,
-      { nome, tipo, valor },
+      { nome, tipo, valor, data },
       { new: true }
     );
+
+    if (!investimentoAtualizado) {
+      return res.status(404).json({ message: 'Investimento não encontrado' });
+    }
+
     res.json(investimentoAtualizado);
   } catch (error) {
     res.status(500).json({ message: 'Erro ao atualizar investimento', error });
@@ -41,9 +54,21 @@ export const updateInvestimento = async (req: Request, res: Response) => {
 
 export const deleteInvestimento = async (req: Request, res: Response) => {
   const { id } = req.params;
+  console.log("ID recebido para exclusão:", id); // Log do ID
+
+   // Valida se o ID é um ObjectId válido
+   if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(400).json({ message: 'ID inválido' });
+  }
+
 
   try {
-    await Investimento.findByIdAndDelete(id);
+    const investimentoExcluido = await Investimento.findByIdAndDelete(id);
+
+    if (!investimentoExcluido) {
+      return res.status(404).json({ message: 'Investimento não encontrado' });
+    }
+
     res.status(204).send();
   } catch (error) {
     res.status(500).json({ message: 'Erro ao excluir investimento', error });
